Heatmap logic and widget transferred to new library
